+++ packages/web3-modal/src/components/modals/AccountModal/index.tsx
@@ -6,12 +6,12 @@ import { useTheme } from 'styled-components'
 
 import { UserOptionsCtrlState } from '../../../controllers/types'
 import { useConnectDisconnect, usePstlWeb3Modal, useUserConnectionInfo } from '../../../hooks'
-import { useConnectedChainAndWalletLogo } from '../../../hooks/misc/useLogos'
+import { useConnectedChainAndWalletIcon } from '../../../hooks/misc/useLogos'
 import { PstlModalTheme } from '../../../theme'
 import { ConnectorEnhanced } from '../../../types'
 import { useDeriveAppType } from '../../../utils/connectors'
 import { BaseModalProps, ModalId } from '../common/types'
-import { WalletChainLogos } from './WalletChainLogos'
+import { WalletChainIcons } from './WalletChainIcons'
 import {
   AccountColumnContainer,
   AccountModalButton,
@@ -46,7 +46,7 @@ function AccountModalContent({ closeModalOnConnect, errorOptions }: PstlAccountM
   })
 
   const appType = useDeriveAppType()
-  const { isUnsupportedChain, showNetworkButton, isNonFrameWalletApp } = useMemo(() => {
+  const { isLedgerHid, isUnsupportedChain, showNetworkButton, isNonFrameWalletApp } = useMemo(() => {
     const isUnsupportedChain = !userConnectionInfo.supportedChains.some(
       (chain) => chain.id === userConnectionInfo.chain?.id
     )
@@ -55,13 +55,14 @@ function AccountModalContent({ closeModalOnConnect, errorOptions }: PstlAccountM
     return {
       isUnsupportedChain,
       showNetworkButton: supportsSeveralChains || isUnsupportedChain,
+      isLedgerHid: userConnectionInfo.connector?.type === 'hid',
       isNonFrameWalletApp: appType === 'DAPP' || appType === 'TEST_FRAMEWORK_IFRAME'
     }
-  }, [appType, userConnectionInfo.chain?.id, userConnectionInfo.supportedChains])
+  }, [appType, userConnectionInfo.chain?.id, userConnectionInfo.connector?.type, userConnectionInfo.supportedChains])
 
   const isSmallerScreen = useIsSmallMediaWidth()
 
-  const { chain: chainLogo, wallet: walletLogo } = useConnectedChainAndWalletLogo()
+  const { chain: chainLogo, wallet: walletLogo } = useConnectedChainAndWalletIcon()
 
   const [isCopied, onCopy] = useCopyClipboard(1500)
   const onExplorer = useCallback(() => {
@@ -77,14 +78,14 @@ function AccountModalContent({ closeModalOnConnect, errorOptions }: PstlAccountM
   return (
     <AccountColumnContainer width="100%" gap="1rem" overflowY="auto">
       {/* Logos Row */}
-      <WalletChainLogos
+      <WalletChainIcons
         wallet={{
           title: userConnectionInfo?.connector?.name || userConnectionInfo?.connector?.id || 'Unknown provider',
-          logo: walletLogo
+          icon: walletLogo
         }}
         chain={{
           title: userConnectionInfo?.chain?.name || userConnectionInfo?.chain?.id?.toString() || 'Unknown chain',
-          logo: chainLogo
+          icon: chainLogo
         }}
       />
       {/* Address and Balance Row */}
@@ -208,6 +209,17 @@ function AccountModalContent({ closeModalOnConnect, errorOptions }: PstlAccountM
                 onClick={() => modalCallbacks.open({ route: 'SelectNetwork', withHistory: true })}
               >
                 Switch Network
+              </AccountModalButton>
+            )}
+            {isLedgerHid && (
+              <AccountModalButton
+                node="main"
+                id={`${ModalId.ACCOUNT}__network-button`}
+                connected={false}
+                padding="0.6rem 1.2rem"
+                onClick={() => modalCallbacks.open({ route: 'HidDeviceOptions', withHistory: true })}
+              >
+                HID Options
               </AccountModalButton>
             )}
             {isNonFrameWalletApp && (