+++ packages/web3-modal/src/components/modals/ConnectionModal/RenderConnectorOptions.tsx
@@ -3,7 +3,7 @@ import React, { Dispatch, SetStateAction } from 'react'
 import { ProviderMountedMap, PstlWeb3ConnectionModalProps } from '.'
 import { UserOptionsCtrlState } from '../../../controllers/types'
 import { useConnectDisconnect, useUserConnectionInfo } from '../../../hooks'
-import { ConnectFunction, ConnectorEnhanced, FullWeb3ModalStore } from '../../../types'
+import { ConfigStore, ConnectFunction, ConnectorEnhanced } from '../../../types'
 import { runConnectorConnectionLogic } from '../../../utils/connectConnector'
 import { ConnectorOption } from './ConnectorOption'
 
@@ -17,7 +17,7 @@ export type RenderConnectorOptionsProps = Pick<
   userConnectionInfo: ReturnType<typeof useUserConnectionInfo>
   connect: ConnectFunction
   disconnect: ReturnType<typeof useConnectDisconnect>['disconnect']['disconnectAsync']
-  modalsStore: FullWeb3ModalStore['ui']
+  modalsStore: ConfigStore['ui']
   providerMountedState: [ProviderMountedMap, Dispatch<SetStateAction<ProviderMountedMap>>]
   providerLoadingState: [boolean, (loading: boolean) => void]
 }
@@ -40,7 +40,7 @@ const RenderConnectorOptionsBase =
     // a. User explicitly states to ignore it
     // b. Window object does NOT contain the injected ethereum proxy object
     if ((hideInjectedFromRoot || (!IS_SERVER && !(window as any)?.ethereum)) && connector.id === 'injected') return null
-    const [{ label, logo, connected, isRecommended }, callback] = runConnectorConnectionLogic(
+    const [{ label, icon, connected, isRecommended }, callback] = runConnectorConnectionLogic(
       connector,
       currentConnector,
       {
@@ -76,7 +76,7 @@ const RenderConnectorOptionsBase =
         label={label}
         isRecommended={isRecommended}
         modalView={modalView}
-        logo={<img src={logo} />}
+        icon={<img src={icon} />}
       />
     )
   }