+++ packages/web3-modal/src/fixtures/config.tsx
@@ -38,11 +38,28 @@ export const pstlModalTheme = createTheme({
           success: '#777b48',
           url: BG_LOGO
         },
+        button: {
+          main: {
+            font: {
+              color: 'ghostwhite'
+            }
+          }
+        },
+        input: {
+          font: {
+            color: 'violet'
+          }
+        },
         tooltip: {
           background: '#685985',
           font: {
             color: 'ghostwhite',
             family: 'monospace'
+          }
+        },
+        text: {
+          main: {
+            color: 'ghostwhite'
           }
         },
         font: {
@@ -82,8 +99,6 @@ export const pstlModalTheme = createTheme({
             },
             border: { border: 'none', radius: '1em' },
             font: {
-              color: 'ghostwhite',
-              size: '1em',
               style: 'normal',
               weight: 200,
               letterSpacing: '-1px',
@@ -115,7 +130,6 @@ export const pstlModalTheme = createTheme({
             filter: 'invert(1) hue-rotate(65deg)'
           },
           alternate: {
-            font: ACCOUNT_BUTTON.font,
             filter: 'hue-rotate(-5deg) contrast(1.3)',
             background: { default: 'indianred', url: 'none' }
           }
@@ -129,26 +143,7 @@ export const pstlModalTheme = createTheme({
           }
         }
       },
-      transactions: {
-        button: {
-          main: {
-            background: {
-              default: '#6c5990'
-            },
-            font: {
-              weight: 100
-            }
-          }
-        },
-        text: {
-          main: {
-            weight: 600
-          },
-          small: {
-            weight: 100
-          }
-        }
-      }
+      transactions: {}
     }
   },
   get DARK() {
@@ -169,42 +164,42 @@ export const pstlModalTheme = createTheme({
 
 export const COMMON_CONNECTOR_OVERRIDES = {
   walletconnect: {
-    logo: WALLETCONNECT_LOGO
+    icon: WALLETCONNECT_LOGO
   },
   web3auth: {
     isRecommended: true,
-    logo: WEB3AUTH_LOGO
+    icon: WEB3AUTH_LOGO
   },
   ledger: {
     customName: 'Ledger Live',
-    logo: 'https://crypto-central.io/library/uploads/Ledger-Logo-3.png',
+    icon: 'https://crypto-central.io/library/uploads/Ledger-Logo-3.png',
     modalNodeId: 'ModalWrapper',
     rank: 0,
     isRecommended: true
   },
   'ledger-hid': {
     customName: 'Ledger HID',
-    logo: 'https://crypto-central.io/library/uploads/Ledger-Logo-3.png',
+    icon: 'https://crypto-central.io/library/uploads/Ledger-Logo-3.png',
     rank: 10,
     isRecommended: true
   },
   'coinbase-wallet-injected': {
     customName: 'Coinbase Wallet',
-    logo: 'https://companieslogo.com/img/orig/COIN-a63dbab3.png?t=1648737284',
+    icon: 'https://companieslogo.com/img/orig/COIN-a63dbab3.png?t=1648737284',
     rank: 12,
     downloadUrl: 'https://www.coinbase.com/wallet/downloads',
     isRecommended: true
   },
   'taho-injected': {
     customName: 'Taho',
-    logo: 'https://user-images.githubusercontent.com/95715502/221033622-fb606b37-93f1-485b-9ce5-59b92f756033.png',
+    icon: 'https://user-images.githubusercontent.com/95715502/221033622-fb606b37-93f1-485b-9ce5-59b92f756033.png',
     rank: 11,
     downloadUrl: 'https://taho.xyz/',
     isRecommended: false
   },
   'metamask-injected': {
     customName: 'MetaMask',
-    logo: 'https://upload.wikimedia.org/wikipedia/commons/thumb/3/36/MetaMask_Fox.svg/1200px-MetaMask_Fox.svg.png',
+    icon: 'https://upload.wikimedia.org/wikipedia/commons/thumb/3/36/MetaMask_Fox.svg/1200px-MetaMask_Fox.svg.png',
     rank: 10,
     downloadUrl: 'https://metamask.io/downloads',
     isRecommended: true