+++ packages/wagmi-connectors/CHANGELOG.md
@@ -1,4 +1,53 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [2.0.2-alpha.3](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/wagmi-connectors@2.0.2-alpha.2...@past3lle/wagmi-connectors@2.0.2-alpha.3) (2024-01-25)
+
+
+### Bug Fixes
+
+* **packages:** update registries to standard NPM uri ([78fa2c8](https://github.com/PAST3LLE/past3lle-monorepo/commit/78fa2c870d2458a22fa0109a2aa29fde94b1cb64))
+
+
+
+
+
+## [2.0.2-alpha.2](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/wagmi-connectors@2.0.2-alpha.1...@past3lle/wagmi-connectors@2.0.2-alpha.2) (2024-01-25)
+
+**Note:** Version bump only for package @past3lle/wagmi-connectors
+
+
+
+
+
+## [2.0.2-alpha.1](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/wagmi-connectors@2.0.2-alpha.0...@past3lle/wagmi-connectors@2.0.2-alpha.1) (2024-01-24)
+
+
+### Bug Fixes
+
+* **workflows, web3-modal, wagmi-connectors:** update workflow scripts,  web3-modal fixes, wagmi-connector add optional params([#57](https://github.com/PAST3LLE/past3lle-monorepo/issues/57)) ([50c59e3](https://github.com/PAST3LLE/past3lle-monorepo/commit/50c59e37d33887d92c8b52399ddc9a7971e688e1))
+
+
+
+
+
+## [2.0.2-alpha.0](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/wagmi-connectors@2.0.0-alpha.3...@past3lle/wagmi-connectors@2.0.2-alpha.0) (2024-01-24)
+
+**Note:** Version bump only for package @past3lle/wagmi-connectors
+
+
+
+
+
 # @past3lle/wagmi-connectors
+
+## 2.0.1
+
+### Patch Changes
+
+- [#51](https://github.com/PAST3LLE/monorepo/pull/51) [`c89e0d6`](https://github.com/PAST3LLE/monorepo/commit/c89e0d68f2bcadfd418e04737b5ba1416d714796) Thanks [@W3stside](https://github.com/W3stside)! - Added version file
 
 ## 2.0.0