+++ packages/web3-modal/CHANGELOG.md
@@ -1,4 +1,90 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [2.1.0-alpha.0](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/web3-modal@2.0.2-alpha.5...@past3lle/web3-modal@2.1.0-alpha.0) (2024-01-25)
+
+
+### Features
+
+* **web3-modal:**  switch network behaviour and hid options in account modal ([#61](https://github.com/PAST3LLE/past3lle-monorepo/issues/61)) ([6bc18e3](https://github.com/PAST3LLE/past3lle-monorepo/commit/6bc18e30f5a53507655fc5702573219d09569a3c))
+
+
+
+
+
+## [2.0.2-alpha.5](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/web3-modal@2.0.2-alpha.4...@past3lle/web3-modal@2.0.2-alpha.5) (2024-01-25)
+
+**Note:** Version bump only for package @past3lle/web3-modal
+
+
+
+
+
+## [2.0.2-alpha.4](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/web3-modal@2.0.2-alpha.3...@past3lle/web3-modal@2.0.2-alpha.4) (2024-01-25)
+
+
+### Bug Fixes
+
+* **packages:** update registries to standard NPM uri ([78fa2c8](https://github.com/PAST3LLE/past3lle-monorepo/commit/78fa2c870d2458a22fa0109a2aa29fde94b1cb64))
+
+
+
+
+
+## [2.0.2-alpha.3](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/web3-modal@2.0.2-alpha.2...@past3lle/web3-modal@2.0.2-alpha.3) (2024-01-25)
+
+**Note:** Version bump only for package @past3lle/web3-modal
+
+
+
+
+
+## [2.0.2-alpha.2](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/web3-modal@2.0.2-alpha.1...@past3lle/web3-modal@2.0.2-alpha.2) (2024-01-24)
+
+
+### Bug Fixes
+
+* **workflows, web3-modal, wagmi-connectors:** update workflow scripts,  web3-modal fixes, wagmi-connector add optional params([#57](https://github.com/PAST3LLE/past3lle-monorepo/issues/57)) ([50c59e3](https://github.com/PAST3LLE/past3lle-monorepo/commit/50c59e37d33887d92c8b52399ddc9a7971e688e1))
+
+
+
+
+
+## [2.0.2-alpha.1](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/web3-modal@2.0.2-alpha.0...@past3lle/web3-modal@2.0.2-alpha.1) (2024-01-24)
+
+
+### Bug Fixes
+
+* **web3-modal:** fix theme missing parts and minor styles ([#56](https://github.com/PAST3LLE/past3lle-monorepo/issues/56)) ([2f247e4](https://github.com/PAST3LLE/past3lle-monorepo/commit/2f247e4fa41da69b827f3f29556481e9e9f0aa0f))
+
+
+
+
+
+## [2.0.2-alpha.0](https://github.com/PAST3LLE/past3lle-monorepo/compare/@past3lle/web3-modal@2.0.0-alpha.12...@past3lle/web3-modal@2.0.2-alpha.0) (2024-01-24)
+
+**Note:** Version bump only for package @past3lle/web3-modal
+
+
+
+
+
 # @past3lle/web3-modal
+
+## 2.0.1
+
+### Patch Changes
+
+- [#51](https://github.com/PAST3LLE/monorepo/pull/51) [`c89e0d6`](https://github.com/PAST3LLE/monorepo/commit/c89e0d68f2bcadfd418e04737b5ba1416d714796) Thanks [@W3stside](https://github.com/W3stside)! - Added version file
+
+- Updated dependencies [[`c89e0d6`](https://github.com/PAST3LLE/monorepo/commit/c89e0d68f2bcadfd418e04737b5ba1416d714796)]:
+  - @past3lle/wagmi-connectors@2.0.1
+  - @past3lle/components@2.0.0
+  - @past3lle/hooks@2.0.0
+  - @past3lle/theme@2.0.1
+  - @past3lle/utils@2.0.1
 
 ## 2.0.0